fix: Removed all new BlockStore implementations; reverted to old directory structure
