+++ wnfs-common/src/lib.rs
@@ -1,5 +1,4 @@
 //! This crate contains the common types and functions used by the WNFS crates.
-
 mod async_serialize;
 pub mod blockstore;
 mod encoding;
